Return an fsHook object
Otherwise we can't actually get to the SetFilePerm method because
obviously it's not in the logger interface and the fsHook interface
isn't exported so we can't cast to it.

This seems to be enough but is this the corect go-y way of doing
this? Should we be exporting the interface anyway?